% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acadYear_from_term}
\alias{acadYear_from_term}
\title{Produce academic year from Peoplesoft Term}
\usage{
acadYear_from_term(term)
}
\arguments{
\item{term}{a peoplesoft term code}
}
\value{
academic year
}
\description{
Produce academic year from Peoplesoft Term
}
