% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{bind_mrows}
\alias{bind_mrows}
\title{bind_mrows - a function to match the types of two identically named data frames
and bind the results.  The first input is also the template, and the second data frame
classes will be changed to match.}
\usage{
bind_mrows(df1 = NULL, df2 = NULL)
}
\arguments{
\item{df1}{First Input data frame. Also the provides the template classes for the result}

\item{df2}{Second input data frame.}
}
\value{
data \code{bind_rows(df1,d2)} with its classes changed to match df1
}
\description{
bind_mrows - a function to match the types of two identically named data frames
and bind the results.  The first input is also the template, and the second data frame
classes will be changed to match.
}

