% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_spaced_offset}
\alias{calc_spaced_offset}
\title{Calculates the vertical offset between successive data points
Credit to James Keirstead for the original function
https://github.com/jkeirstead/r-slopegraph}
\usage{
calc_spaced_offset(.data, min_space, target)
}
\arguments{
\item{.data}{a data frame representing a single year of data}

\item{min_space}{the minimum spacing between y values}

\item{target}{target column to sort}
}
\value{
a data frame
}
\description{
Calculates the vertical offset between successive data points
Credit to James Keirstead for the original function
https://github.com/jkeirstead/r-slopegraph
}

