% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_from_term}
\alias{date_from_term}
\title{Produce Date from Peoplesoft Term}
\usage{
date_from_term(term)
}
\arguments{
\item{term}{a peoplesoft term code}
}
\value{
a useable date
}
\description{
Produce Date from Peoplesoft Term
}
