% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_frozen_data}
\alias{fix_frozen_data}
\title{Fix peoplesoft frozen data, accounting for degree, program and plan changes}
\usage{
fix_frozen_data(x)
}
\arguments{
\item{x}{input PS dataframe}
}
\value{
x the cleaned dataframe
}
\description{
Fix peoplesoft frozen data, accounting for degree, program and plan changes
}
