% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ga_pal}
\alias{ga_pal}
\title{Graduate Attribute Colors}
\format{An object of class \code{character} of length 12.}
\usage{
ga_pal
}
\description{
Graduate Attribute Colors
}
\keyword{datasets}
