% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{wrapper for Not In.  Finds values that don't match to a provided vector}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{input vector to compare}

\item{table}{values to compare against}
}
\description{
wrapper for Not In.  Finds values that don't match to a provided vector
}

