% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{greenfocus}
\alias{greenfocus}
\title{Green as highlight}
\format{
An object of class \code{character} of length 8.
}
\usage{
greenfocus
}
\description{
Green as highlight
}
\keyword{datasets}
