% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brightspace.R
\name{import_brightspace}
\alias{import_brightspace}
\title{import_brightspace}
\usage{
import_brightspace(df, mode = "median")
}
\arguments{
\item{df}{a data frame in brightspace export wide-format}
}
\value{
df
}
\description{
This function takes an input data frame, in D2L export format and exports it
 into a tidy data frame in a specific tidy format
}

