% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_FEAS}
\alias{in_FEAS}
\title{conveinence function to quickly filter all plans/concentrations under engineering
currently used only for undergrad}
\usage{
in_FEAS(x)
}
\arguments{
\item{x}{academic plan or concetration}
}
\description{
conveinence function to quickly filter all plans/concentrations under engineering
currently used only for undergrad
}
