% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inFEAS}
\alias{inFEAS}
\title{conveinence function to quickly filter all plans/concentrations under engineering
currently used only for undergrad}
\usage{
inFEAS(x)
}
\arguments{
\item{x}{academic plan or concetration}
}
\description{
conveinence function to quickly filter all plans/concentrations under engineering
currently used only for undergrad
}
