% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{install_github_packages}
\alias{install_github_packages}
\title{Install github packages by repository name}
\usage{
install_github_packages(package, force = FALSE)
}
\arguments{
\item{package}{name of the package to install e.g "ggplot2"}

\item{force}{force re-installation of the package, defults to FALSE}
}
\description{
To Do: - build ability to take a list/vector or single input
       - figure out how to wrap errors properly
       - figure out how to return errors properly
       - figure out how to install only R code packages (html widgets list as html)
       - figure out better means of searching github
}
\examples{
install_github_packages("ggplot2")
}

