% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{label_wrap_md_gen}
\alias{label_wrap_md_gen}
\title{label_wrap_md_gen}
\usage{
label_wrap_md_gen(width = 25, multi_line = TRUE)
}
\arguments{
\item{width}{line with}

\item{multi_line}{is it multiline}
}
\description{
label_wrap_md_gen
}
