% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{lumina_pal}
\alias{lumina_pal}
\title{Lumina color palette}
\format{An object of class \code{character} of length 5.}
\usage{
lumina_pal
}
\description{
Lumina color palette
}
\keyword{datasets}
