% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{match_type}
\alias{match_type}
\title{match_type - a function to match the types of two identically named data frames}
\usage{
match_type(data = NULL, template = NULL)
}
\arguments{
\item{data}{data frame whose classes are to be changed}

\item{template}{data frame providing the template for the correct data classes}
}
\value{
data data with its classes changed to match df1
}
\description{
match_type - a function to match the types of two identically named data frames
}

