% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{mountain_pal}
\alias{mountain_pal}
\title{Harris Mountain Form color palette}
\format{An object of class \code{character} of length 5.}
\usage{
mountain_pal
}
\description{
Harris Mountain Form color palette
}
\keyword{datasets}
