% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{p_key_check}
\alias{p_key_check}
\title{Function to check a data frame for primary key integrity prior to
copying to the database.}
\usage{
p_key_check(df, p_key)
}
\arguments{
\item{df}{Data frame to check.  Columns must contain all elements of the key to test}

\item{p_key}{Character vector of colnames that comprise the primary key}
}
\description{
Function to check a data frame for primary key integrity prior to
copying to the database.
}

