% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_ps_year}
\alias{scale_ps_year}
\title{Custom formatter for peoplesoft terms}
\usage{
scale_ps_year(labels, abbr = FALSE)
}
\arguments{
\item{labels}{vector of labels}

\item{abbr}{logical, abbreviate axis}
}
\value{
formatted year labels
}
\description{
Custom formatter for peoplesoft terms
}
