% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_term_Year}
\alias{scale_term_Year}
\title{Custom formatter for peoplesoft terms}
\usage{
scale_term_Year(labels)
}
\arguments{
\item{labels}{vector of labels}
}
\value{
formatted year labels
}
\description{
Custom formatter for peoplesoft terms
}
