% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spaced_title}
\alias{spaced_title}
\title{Make word into W O R D}
\usage{
spaced_title(text)
}
\arguments{
\item{text}{word to space and capitalize}
}
\value{
T E X T
}
\description{
Make word into W O R D
}
