% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_break}
\alias{str_break}
\title{str_break}
\usage{
str_break(string, width = 80, indent = 0, exdent = 0)
}
\arguments{
\item{string}{string to break}

\item{width}{width to break at}

\item{indent}{left indent}

\item{exdent}{right indent}
}
\value{
string with line braaks at specified width
}
\description{
str_break
}
