% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tol11qualitative}
\alias{tol11qualitative}
\title{Qualitative color schemes by Paul Tol}
\format{
An object of class \code{character} of length 11.
}
\usage{
tol11qualitative
}
\description{
Qualitative color schemes by Paul Tol
}
\keyword{datasets}
