% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tol1qualitative}
\alias{tol1qualitative}
\title{Qualitative color schemes by Paul Tol}
\format{An object of class \code{character} of length 1.}
\usage{
tol1qualitative
}
\description{
Qualitative color schemes by Paul Tol
}
\keyword{datasets}
