% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tricolour_pal}
\alias{tricolour_pal}
\title{Queen's tri-color palette}
\format{
An object of class \code{character} of length 3.
}
\usage{
tricolour_pal
}
\description{
Queen's tri-color palette
}
\keyword{datasets}
