% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{tt_dir}
\alias{tt_dir}
\title{Create a new TidyTuesday directory}
\usage{
tt_dir(
  year = lubridate::year(Sys.Date()),
  week = lubridate::isoweek(Sys.Date()) + 1
)
}
\arguments{
\item{year}{4 digit year}

\item{week}{numeric week}
}
\description{
Create a new TidyTuesday directory
}
