% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spaced_sort}
\alias{spaced_sort}
\title{Spaced sort for slopegraphs}
\usage{
spaced_sort(.data, group, target, min_space = 0.05)
}
\arguments{
\item{.data}{the raw data frame}

\item{group}{grouping column to split data on}

\item{target}{target column to sort}

\item{min_space}{fraction of total data range to leave as a
minium gap}
}
\value{
a data frame with the ypos column added
}
\description{
Calculates the position of each element to ensure a minimum
space between adjacent entries within a column, while preserving
rank order.  Group lines can cross
}
\details{
Credit to James Keirstead for the original function
https://github.com/jkeirstead/r-slopegraph
}

